HODLR_TP_sample_ls

Assuming a tensor product \(f*g\), this function samples the length-scale for f, given the marginal distribution of Y given g. It performs a simple Metropolis step where the proposal distribution is a uniform(-0.5,0.5) and the prior distribution on the length-scale is a uniform distribution specified by the bounds argument.

HODLR_TP_sample_ls(
  ls,
  Y,
  TP,
  x,
  cur_tau,
  bounds,
  idx = NULL,
  kernel = "matern"
)

Arguments

ls

The current value of the length scale being used for f.

Y

The observed data or the residual given other components in the mean.

TP

This parameter represents g. In practice, this value can represent multiple dimensions.

x

Values of the covariate associated with f.

idx

This optional parameter is the index of the x sorted in increasing order. If it is not specified, the function will sort x for the Gibbs sampler. However, since most samplers only require x to be sorted once, this is not efficient. We recommend passing the index (i.e. order(x)).

kernel

The kernel function to use. By default the "matern" kernel is used but the squared exponential kernel "sqr-exp" is also available.

Examples